+++ src/Twilio/Rest/FlexApi/V1.php
@@ -15,7 +15,9 @@ use Twilio\InstanceContext;
 use Twilio\Rest\FlexApi\V1\ChannelList;
 use Twilio\Rest\FlexApi\V1\ConfigurationList;
 use Twilio\Rest\FlexApi\V1\FlexFlowList;
+use Twilio\Rest\FlexApi\V1\GoodDataList;
 use Twilio\Rest\FlexApi\V1\InteractionList;
+use Twilio\Rest\FlexApi\V1\UserRolesList;
 use Twilio\Rest\FlexApi\V1\WebChannelList;
 use Twilio\Version;
 
@@ -23,7 +25,9 @@ use Twilio\Version;
  * @property ChannelList $channel
  * @property ConfigurationList $configuration
  * @property FlexFlowList $flexFlow
+ * @property GoodDataList $goodData
  * @property InteractionList $interaction
+ * @property UserRolesList $userRoles
  * @property WebChannelList $webChannel
  * @method \Twilio\Rest\FlexApi\V1\ChannelContext channel(string $sid)
  * @method \Twilio\Rest\FlexApi\V1\FlexFlowContext flexFlow(string $sid)
@@ -34,7 +38,9 @@ class V1 extends Version {
     protected $_channel;
     protected $_configuration;
     protected $_flexFlow;
+    protected $_goodData;
     protected $_interaction;
+    protected $_userRoles;
     protected $_webChannel;
 
     /**
@@ -68,11 +74,25 @@ class V1 extends Version {
         return $this->_flexFlow;
     }
 
+    protected function getGoodData(): GoodDataList {
+        if (!$this->_goodData) {
+            $this->_goodData = new GoodDataList($this);
+        }
+        return $this->_goodData;
+    }
+
     protected function getInteraction(): InteractionList {
         if (!$this->_interaction) {
             $this->_interaction = new InteractionList($this);
         }
         return $this->_interaction;
+    }
+
+    protected function getUserRoles(): UserRolesList {
+        if (!$this->_userRoles) {
+            $this->_userRoles = new UserRolesList($this);
+        }
+        return $this->_userRoles;
     }
 
     protected function getWebChannel(): WebChannelList {