+++ src/Twilio/Rest/Messaging/V1/Service/UsAppToPersonList.php
@@ -11,6 +11,7 @@ namespace Twilio\Rest\Messaging\V1\Service;
 
 use Twilio\Exceptions\TwilioException;
 use Twilio\ListResource;
+use Twilio\Options;
 use Twilio\Serialize;
 use Twilio\Stream;
 use Twilio\Values;
@@ -47,10 +48,13 @@ class UsAppToPersonList extends ListResource {
      *                               messages that contain links
      * @param bool $hasEmbeddedPhone Indicates that this SMS campaign will send
      *                               messages that contain phone numbers
+     * @param array|Options $options Optional Arguments
      * @return UsAppToPersonInstance Created UsAppToPersonInstance
      * @throws TwilioException When an HTTP error occurs.
      */
-    public function create(string $brandRegistrationSid, string $description, array $messageSamples, string $usAppToPersonUsecase, bool $hasEmbeddedLinks, bool $hasEmbeddedPhone): UsAppToPersonInstance {
+    public function create(string $brandRegistrationSid, string $description, array $messageSamples, string $usAppToPersonUsecase, bool $hasEmbeddedLinks, bool $hasEmbeddedPhone, array $options = []): UsAppToPersonInstance {
+        $options = new Values($options);
+
         $data = Values::of([
             'BrandRegistrationSid' => $brandRegistrationSid,
             'Description' => $description,
@@ -58,6 +62,13 @@ class UsAppToPersonList extends ListResource {
             'UsAppToPersonUsecase' => $usAppToPersonUsecase,
             'HasEmbeddedLinks' => Serialize::booleanToString($hasEmbeddedLinks),
             'HasEmbeddedPhone' => Serialize::booleanToString($hasEmbeddedPhone),
+            'MessageFlow' => $options['messageFlow'],
+            'OptInMessage' => $options['optInMessage'],
+            'OptOutMessage' => $options['optOutMessage'],
+            'HelpMessage' => $options['helpMessage'],
+            'OptInKeywords' => Serialize::map($options['optInKeywords'], function($e) { return $e; }),
+            'OptOutKeywords' => Serialize::map($options['optOutKeywords'], function($e) { return $e; }),
+            'HelpKeywords' => Serialize::map($options['helpKeywords'], function($e) { return $e; }),
         ]);
 
         $payload = $this->version->create('POST', $this->uri, [], $data);