+++ tests/Twilio/Integration/Messaging/V1/Service/UsAppToPersonTest.php
@@ -72,6 +72,19 @@ class UsAppToPersonTest extends HolodeckTestCase {
                         "brand_tier": "TOP"
                     }
                 },
+                "message_flow": "End users opt-in by visiting www.examplewebsite.com and adding their phone number. They then check a box agreeing to receive text messages from Example Brand. Additionally, end users can also opt-in by texting START to (111) 222-3333 to opt in.",
+                "opt_in_message": "Acme Corporation: You are now opted-in. For help, reply HELP. To opt-out, reply STOP",
+                "opt_out_message": "You have successfully been unsubscribed from Acme Corporation. You will not receive any more messages from this number.",
+                "help_message": "Acme Corporation: Please visit www.acmecorporation.com to get support. To opt-out, reply STOP.",
+                "opt_in_keywords": [
+                    "START"
+                ],
+                "opt_out_keywords": [
+                    "STOP"
+                ],
+                "help_keywords": [
+                    "HELP"
+                ],
                 "date_created": "2021-02-18T14:48:52Z",
                 "date_updated": "2021-02-18T14:48:52Z",
                 "url": "https://messaging.twilio.com/v1/Services/MGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Compliance/Usa2p/QE2c6890da8086d771620e9b13fadeba0b",
@@ -159,6 +172,19 @@ class UsAppToPersonTest extends HolodeckTestCase {
                                 "brand_tier": "TOP"
                             }
                         },
+                        "message_flow": "End users opt-in by visiting www.examplewebsite.com and adding their phone number. They then check a box agreeing to receive text messages from Example Brand. Additionally, end users can also opt-in by texting START to (111) 222-3333 to opt in.",
+                        "opt_in_message": "Acme Corporation: You are now opted-in. For help, reply HELP. To opt-out, reply STOP",
+                        "opt_out_message": "You have successfully been unsubscribed from Acme Corporation. You will not receive any more messages from this number.",
+                        "help_message": "Acme Corporation: Please visit www.acmecorporation.com to get support. To opt-out, reply STOP.",
+                        "opt_in_keywords": [
+                            "START"
+                        ],
+                        "opt_out_keywords": [
+                            "STOP"
+                        ],
+                        "help_keywords": [
+                            "HELP"
+                        ],
                         "date_created": "2021-02-18T14:48:52Z",
                         "date_updated": "2021-02-18T14:48:52Z",
                         "url": "https://messaging.twilio.com/v1/Services/MGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Compliance/Usa2p/QE2c6890da8086d771620e9b13fadeba0b",
@@ -228,6 +254,19 @@ class UsAppToPersonTest extends HolodeckTestCase {
                         "brand_tier": "TOP"
                     }
                 },
+                "message_flow": "End users opt-in by visiting www.examplewebsite.com and adding their phone number. They then check a box agreeing to receive text messages from Example Brand. Additionally, end users can also opt-in by texting START to (111) 222-3333 to opt in.",
+                "opt_in_message": "Acme Corporation: You are now opted-in. For help, reply HELP. To opt-out, reply STOP",
+                "opt_out_message": "You have successfully been unsubscribed from Acme Corporation. You will not receive any more messages from this number.",
+                "help_message": "Acme Corporation: Please visit www.acmecorporation.com to get support. To opt-out, reply STOP.",
+                "opt_in_keywords": [
+                    "START"
+                ],
+                "opt_out_keywords": [
+                    "STOP"
+                ],
+                "help_keywords": [
+                    "HELP"
+                ],
                 "date_created": "2021-02-18T14:48:52Z",
                 "date_updated": "2021-02-18T14:48:52Z",
                 "url": "https://messaging.twilio.com/v1/Services/MGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Compliance/Usa2p/QE2c6890da8086d771620e9b13fadeba0b",