+++ CHANGES.md
@@ -1,6 +1,18 @@
 twilio-php Changelog
 ====================
 
+[2022-11-10] Version 6.43.2
+---------------------------
+**Flex**
+- Adding two new authorization API 'user_roles' and 'gooddata' in version `v1`
+
+**Messaging**
+- Add new Campaign properties (MessageFlow, OptInMessage, OptInKeywords, OptOutMessage, OptOutKeywords, HelpMessage, HelpKeywords)
+
+**Twiml**
+- Add new speech models to `Gather`.
+
+
 [2022-10-31] Version 6.43.1
 ---------------------------
 **Api**